refactor: extract GenerationOrchestrator from GenerateCommand
Extract core orchestration logic into a dedicated GenerationOrchestrator
class with IGenerationOrchestrator interface. GenerateCommand.ExecuteAsync
now delegates to the orchestrator, reducing from ~130 lines to ~30 lines.

- Create IGenerationOrchestrator interface with RunAsync() method
- Create GenerationOrchestrator implementing the full generation pipeline
- Make IGenerationReporter and IMultiFileOutputReport public for wider use
- Remove WriteSingleFile, WriteMultipleFiles, FormatFileSize, ShowWarnings,
  ValidateOpenApiSpec from GenerateCommand (moved to orchestrator)
- All 1958 existing tests pass without modification
